horridly closes #7: a fixed-position DIV
Added a fixed-position DIV in the left bottom corner.
I haven't known a way to handle this with `contextMenu` or
`brwoserAction`.
For pages with frames or iframes, the DIV still appears, but works only
for those having the same domain with the left-bottom frame (or `top` if
no frames but iframes). Frames and iframes with different
`document.domain` would have warnings (luckily not error) in console.

other modifications:
* skip rendered pieces by recognizing their Element#className by RegExp
`/(^| )LER(-| |$)/`
* add tags "OPTION", "BUTTON" to be skipped
* `lyIDs.js` had wrong name for 90077(外交部特派員公署組織條例); see also
g0v/laweasyread-data/issues/6
* add some `exclude_matches` in `manifest.json`