resolves #14 and support multi tabs in popup
`addIFrame` is discarded; a brand new `addPopup` functions very
differently.
Instead of appending iframes as siblings of links, MouseOver event
(assigned during parsing) now triggers a DIV element to show up, which
is appended as the last child of `document.body` and positioned by
`MouseEvent`.
It must be much easier to satisfy #21 and show other information hard to
display as iframes.